requirements.txt
# API e servidor
fastapi
uvicorn[standard]
gunicorn

# PostgreSQL (sync + async)
psycopg2-binary
asyncpg
sqlalchemy
databases[postgresql]

# Cache Redis
redis
aioredis

# Parsing, ingestão e manipulação de dados
pandas
pyarrow
openpyxl

# Utilidades e CLI
tqdm
python-dotenv
loguru
rich
tabulate

# Web, scraping e requisições (se necessário)
httpx
aiohttp
beautifulsoup4
lxml

# Serialização/validação extra
pydantic
pydantic-settings
PyYAML

# Segurança e rate limiting (opcional)
slowapi

python-dotenv
bcrypt
pyjwt
aiosqlite
